requirements.txt
openai-whisper
transformers
sentencepiece
torch
fastapi
uvicorn
python-multipart
soundfile
ffmpeg-python
gradio
evaluate
jiwer
rouge_score
fasttext-wheel
protobuf
